release 1.1

+++ app/src/main/java/com/mobile/bmg/model/api/ResponseOrganizationTotals.java
@@ -14,5 +14,5 @@ public class ResponseOrganizationTotals {
     public double donationAmount;
 
     @SerializedName("donation_count")
-    public double count;
+    public int count;
 }
